+++ src/main/kotlin/com/greenkode/trader/portfolio/Positions.kt
@@ -1,10 +1,9 @@
 package com.greenkode.trader.portfolio
 
 import com.greenkode.trader.domain.Symbol
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 open class Positions(
     open var timestamp: LocalDateTime? = null,
-    open val positions: MutableMap<Symbol, BigDecimal>
+    open val positions: MutableMap<Symbol, Double>
 )
